+++ nautilus_trader/model/book.pyx
@@ -243,8 +243,16 @@ cdef class OrderBook(Data):
         sequence : uint64, default 0
             The unique sequence number for the update. If default 0 then will increment the `sequence`.
 
+        Raises
+        ------
+        RuntimeError
+            If the book type is L1_MBP.
+
         """
         Condition.not_none(order, "order")
+
+        if self.book_type == BookType.L1_MBP:
+            raise RuntimeError("Invalid book operation: cannot add order for L1_MBP book")
 
         orderbook_add(&self._mem, order._mem, ts_event, sequence)